# src/main/java/com/example/bankcards/dto/CardReadDto.java
package com.example.bankcards.dto;

import com.example.bankcards.entity.Status;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;
import java.time.LocalDate;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class CardReadDto {

    Long id;
    String maskedNumber;
    LocalDate expiredDate;
    Status status;
    BigDecimal balance;
}